extract logic into new HiddenRequirement component
